Improve bindings and binding guidelines
