+++ crates/ctt-bc7enc-rdo/src/bindings.rs
