--- crates/ctt-bc7enc-rdo/src/bindings/mod.rs
@@ -1,1 +1,0 @@
-pub mod bc7e;
